+++ examples/tabs.rs
@@ -84,6 +84,7 @@ fn main() {
                                             serv_name: &serv_name,
                                             chan_name: &chan_name,
                                         },
+                                        false,
                                     );
                                 }